feat: write gatsby-node with TypeScript
