+++ gatsby-theme-terminal/package.json
@@ -31,7 +31,8 @@
     "prism-react-renderer": "^1.2.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0"
+    "react-helmet": "^6.1.0",
+    "ts-node": "^10.3.0"
   },
   "devDependencies": {
     "@types/mdx-js__react": "^1.5.4",
