+++ gatsby-theme-terminal/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "incremental": true,
-    "target": "es5",
+    "target": "es2020",
     "jsx": "preserve",
     "module": "commonjs",
     "rootDir": "./",
